Use git submodule to deduplicate demo code.
It seems more managable than duplicated code, but it is not yet ideal
either. Some subideal aspects:

The demo on the website currently has the header and footer of the main
page; this is now removed. Is that an issue?

It now puts the demo at /content/demo, but for this has to put style.css
into /content/. We could consider to instead put the whole /web/dist
from upstream into /content/web (or whatever name), but then the demo
URL will change to be web/demo.

Files get removed, built and copied in a single line in package.json.
Better suggestions?

One needs to keep the submodule in sync; may be okay, updating the demo
is a manual process already anyway.