Add 11ty for docs content
The Markdown and templates (see `_layouts`) live in
the `src/` folder. All other static content (without
any need for processing) live directly in the `content/`
folder. Eleventy will build what is in `src/` and overwrite
the equivalents in `content/`--it will not delete anything,
however.

The `content/` folder should still be committed to the repo
because we serve out of what is on disk immediately after
checkout from the `content/` folder.